hook for section panel detail
