+++ frontend/components/ResultsPage/Results/DesktopSectionPanel.tsx
@@ -2,6 +2,7 @@ import React from 'react'
 import Section from '../../classModels/Section'
 import WeekdayBoxes from './WeekdayBoxes'
 import NotifCheckBox from '../../panels/NotifCheckBox'
+import useSectionPanelDetail from './useSectionPanelDetail';
 
 interface DesktopSectionPanelProps {
   section: Section
@@ -11,27 +12,8 @@ interface DesktopSectionPanelProps {
 
 
 function DesktopSectionPanel({ section, showNotificationSwitches } : DesktopSectionPanelProps) {
-  const renderTimes = () => {
-    return section.meetings.map((meeting) => {
-      return meeting.times.map((time) => (
-        <>
-          <span>
-            {`${time.start.format('h:mm')}-${time.end.format('h:mm a')} | ${meeting.getLocation()}`}
-          </span>
-          <br />
-        </>
-      ))
-    })
-  }
-  const getSeatsClass = () => {
-    const seatingPercentage = section.seatsRemaining / section.seatsCapacity
-    if (seatingPercentage > (2 / 3)) {
-      return 'green'
-    } if (seatingPercentage > (1 / 3)) {
-      return 'yellow'
-    }
-    return 'red'
-  }
+
+  const {renderTimes, getSeatsClass} = useSectionPanelDetail(section)
 
   return (
     <tr className='DesktopSectionPanel' key={ section.getHash() }>
